api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==0.28.1
python-dotenv==1.0.0
requests==2.31.0
tiktoken==0.5.2 